Refactored the loise-img module.
loise-img was brought into line with the more recently-created (and
well-thought-out) module loise-ascii. In particular, options are passed in each
function, reducing the arity and need for multiplicity of functions.

This addresses the rest of issue #1. (The other part of issue #1 was addressed
with commit 95bd1ee5d5cb63870c952db95dd97df7c4769ff7 and done as part of issue